JVMCBC-232: Rename SystemMetricsEvent to RuntimeMetricsEvent
Motivation
----------
While writing the docs for this feature, it became visible that the
name was poorly chosen because we have a system event type and the
SystemMetricsEvent is a metric event type. All very confusing.

Modifications
-------------
The SystemMetricsEvent has been renamed to RuntimeMetricsEvent to
avoid the "system" name confusion and also to more accurately highlight
from which component the metrics are gathered.

Result
------
No more confusions, hopefully.

Change-Id: I6aae1edcf846875935fd454632a49b6c96f0fe62
Reviewed-on: http://review.couchbase.org/54573
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>