JVMCBC-232: Hide the builder from the DefaultMetricsCollectorConfig.
Motivation
----------
Because the builder just exposes the emit time and emit time unit, the
same thing can easily be done from a create() factory method. In fact,
exposing the builder just makes things complicated.

Modifications
-------------
The builder on the config has been made private but is still used
internally, so if we add more settings we can just make it public
again and expose all the complex options we want.

Result
------
Less methods a user can call and therefore making it less complex
to work with the config.

Change-Id: Ibe2f7b72e5756b278026770c37c31fabdf3fdb67
Reviewed-on: http://review.couchbase.org/54574
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>