JVMCBC-232: Always log metric events.
Motivation
----------
To actually use the metrics we collect in an "always on" fashion,
th default implementation needs to consume metrics and log them.

Modifications
-------------
The environment spins up the logger, listens on metric events and
logs them in a JSON format. The user can disable this completely
or modify the settings of the logger.

Result
------
The metrics will be logged on the default interval unless turned
off by the user explicitly.

Change-Id: Ie8a5923cb88c3ac38e9c2f6f2fdb3454f6a65168
Reviewed-on: http://review.couchbase.org/54585
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>