JVMCBC-235: Optimize ServiceRegistry cache.
Motivation
----------
On every flush, the node needs to loop through all the services and
send down the flush signal. With the previous implementation (the
list), an iterator has been created all the time which employs
unnecessary overhead.

Modifications
-------------
The service registry API is changed to return an array for the service
cache instead of a list which makes it much faster to iterate over,
and since it is heavy read and no writes during traffic, there is no need
for a dynamic datastructure.

Iterating is now done without an iterator, and a little faster than over
an ArrayList.

Result
------
Faster service cache and less allocations on the hot code path. Mind you,
its a small change, but lots of those will add up as well.

Change-Id: Ic29b5bb37016a34777493cca579f19134279d7a0
Reviewed-on: http://review.couchbase.org/54817
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>