JVMCBC-237: Poll for new config every 10 seconds.
Motivation
----------
When no traffic is going through the system and under various
other conditions, without polling for a new config on a regular
basis, there is a chance that we miss the addition of query only
service nodes.

Modifications
-------------
This changeset consists of two changes:

	- The CarrierRefresher now issues a signalOutdated every
	  10 seconds, which in turn will cause all open buckets
	  to fetch new configs, but only for carrier pub. If
	  the HttpRefresher is used, this technique is not used
	  since it's not needed.
	- The KeyValueHandler now does not piggyback a new config
	  loading on every heartbeat, because this is now subsumed
	  by the global 10 second poll interval.

Result
------
It is guaranteed that after 10 seconds we get a new config, as a
result detecting any topology changes which might have fallen
through the cracks until this point.

Note that query service removals are detected much quicker when the
socket closes, a new config is also fetched proactively.

Change-Id: I6afe4911a70d65786d1ef90b3983e6c2303ed32f
Reviewed-on: http://review.couchbase.org/54875
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>