JVMCBC-238: Convert HTTP 500 into a specific ResponseStatus
Motivation
----------
Sometimes the view engine reports with a 500 response in a failure
setting and we should properly turn that into a non-generic response
status.

Modifications
-------------
HTTP 500 -> ResponseStatus.INTERNAL_ERROR, since it is an internal
server error.

Note that there is no impact on view retry handling whatsoever, because
this part is only concerned about the numeric http status code that of
course is not changed.

Results
-------
Properly convert the http 500 response status.

Change-Id: Ie970acd21e6cf273185a5285e060516dc271edf3
Reviewed-on: http://review.couchbase.org/54919
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>