JVMCBC-245: Add time unit to the latency metric.
Motivation
----------
The metric should also include the target unit, so later inspection
of the resulting data is not dependent on identifying the time unit
set at configuration time.

Modifications
-------------
Adding the time unit to all possible output formats.

Result
------
Easier identification of the time unit reported in the latency metrics.

Change-Id: Ifd4b97005a0ba1871e392c6aa7f586c821c8ac98
Reviewed-on: http://review.couchbase.org/55394
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>