JVMCBC-248: Properly retry SSL Handshake failures
Motivation
----------
SSL Handshake failures did bubble up the IO exception handling,
but were not reported and "handled" properly inside the couchbase
code.

Modifications
-------------
Because SSL handshake is a vital part of the connection process,
it makes sense to not complete the connection phase if SSL
handshake fails. In order to do this properly, the original
future is intercepted and proactively failed if a SSL handshake
exception bubbles up to the error handler.

That way, we can still keep the retry logic in a centralized place
and react properly.

Result
------
The code now proactively checks for ssl handshake failures and
reacts accordingly.

Change-Id: I6d45430c9eb5823998121ca5931304dc585a0ab7
Reviewed-on: http://review.couchbase.org/55662
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>