JVMCBC-250: Explicitly set socket timeout and handle it if needed.
Motivation
----------
A socket timeout needs to be set to something lower than the netty
default (30 seconds), since especially in the 5 second bucket open
timeout case, this is too long.

This has some side effects where the code can wait on the connect
phase longer than needed, leading to "infinite" open sequences
that would've been able to go to a different node instead.

Modifications
-------------
Apply a 1 second socket timeout and allow it to be configurable
as well. In addition, the code now handles the netty-specific
exception explicitly to terminate an incoming observable more
quickly, allowing the loader code up the stack to fallback
to another node much more quickly.

Result
------
More deterministic behaviour if a socket cannot be opened.

Change-Id: Ie498389717e6b013a7071b3c88040a4c54e963e8
Reviewed-on: http://review.couchbase.org/55698
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-on: http://review.couchbase.org/55731
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>