JCBC-850: Initial API for Full Text.
Motivation
----------
The Couchbase Server "Watson" release will add support for
full text indexing. As a result, the SDKs need to provide
first-class APIs to make it possible to easily query those
indexes.

Modifications
-------------
This changeset introduces a protocol support for CBFT as well as service
discovery. It also exposes commands for index query and management. CBFT
that is subject to change, but will provide users a straightforward way
to try it out and give feedback in the developer preview and beta
timeframe.

Result
------
First experimental CBFT API.

Change-Id: I94bc71ffab031a91e9133524ae23e92fe9065dcd
Reviewed-on: http://review.couchbase.org/55844
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>