JVMCBC-252: Reduce GC on KV ResponseStatus Matching
Motivation
----------
Looking innocent, the values() iterator on the enum actually produces
quite some GC since it always converts the enum into an array which
needs to be collected a little later.

Modifications
-------------
We only need a handful of codes 99% of the time, so the code is
optimized in if else/if clauses right at the top. If its something
different, we fall back into the loop.

Result
------
For the common cases, no GC created and its also quicker.

Change-Id: If435c06d5bce7a64701f023e3f21eba5fd78374b
Reviewed-on: http://review.couchbase.org/56001
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>