JVMCBC-254: Refactor Ketama Lookup into Config
Motivation
----------
In preparation of a planned refactor with different, more optimized
data structures, the code needs to be better encapsulated.

Modifications
-------------
The actual locating code is now placed inside the configuration, so
the internal details are encapsulated. The proper method exposed
just asks for the id and returns the InetAddress which in turn is
then used in the locator to match with the node list.

Result
------
This is just a refactoring, no behaviour has been changed along
the way. Code is now prepared for internal data structure
modifications.

Change-Id: Iae5d5a7c4dcdfdd5df7d537d4d5fd536d59876ac
Reviewed-on: http://review.couchbase.org/56119
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>