JVMCBC-256: Provide migration path for sync env shutdown.
Motivation
----------
Against the principle of least surprise, the shutdown method on the environment is
asynchronous. Since the environment is also exposed to sync APIs without an async
counterpart, this is unintuitive.

The environment needs to provide sync and async methods for shutdown to cater all
kinds of use cases.

Modifications
-------------
It has been decided that in 2.3.0 we'll make the shutdown method synchronous. To
help people right now, the current shutdown method is deprecated and a new
shutdownAsync method has been added. The old method has now a deprecation comment
to guide users to the right method, avoid breaking their code in effect.

Most people want to block at the shutdown anyways, so the breakage will be very
minimal, if at all (most users ignore the response of shutdown and just "call"
it).

Result
------
Smoother migration path towards a better API on environment shutdown.

Change-Id: Ica7cec3a46f46413e5baf9eedffb060ddc4a0892
Reviewed-on: http://review.couchbase.org/56190
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>