JVMCBC-217: Flow control for DCP
Motivation
----------
Server supports a mechanism to track consumer buffer sizes and stepback
if it sees that the client lacks free spaces in his buffer. This feature
is useful to rate limit the stream of changes on high load producers.

Modification
------------
Implement tracking of consumed bytes, and expose settings to control
size of connection buffer, and the threshold when to send buffer
acknowledgement message.

Result
------
Rate limiting DCP streams is possible.

Change-Id: Idedad59c19c220ebbea21f15006fa7b23aa4dd08
Reviewed-on: http://review.couchbase.org/56553
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>