JVMCBC-262: CarrierRefresher needs to skip non-kv nodes.
Motivation
----------
In a MDS scenario, the refresher previously tried to talk
to any node in the cluster list, even if it doesn't have
a KV service enabled.

Modifications
-------------
The refresher code now checks if the node actually has the
BINARY service enabled, and only in this case it will try
to send a config request to it.

Unit tests have been updated.

Result
------
In an MDS scenario the client now doesn't try to load a
config over CCCP from a node where no KV service is
enabled.

Change-Id: I27711bd4f5bc02b6305aa6ed3db6a2b9eb087f75
Reviewed-on: http://review.couchbase.org/56639
Reviewed-by: Simon Baslé <simon@couchbase.com>
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-on: http://review.couchbase.org/56695