JVMCBC-263: Fix view row parsing failing on } in id/key/value
Motivation:
-----------
The ViewHandler would not detect a closing brackets inside a String as
irrelevant and would stop parsing the row at that point, later on making
Jackson unmarshalling of the row fail.

That is because the ByteProcessor wasn't configured to detect string
boundaries.

Modification:
-------------
Made the ViewHandler use of ClosingPositionBufProcessor correct by
instructing to take string boundaries escaping into account.

Added a unit test case with a closing bracket in id, value and keys
sections of the view response to validate behavior.

Result:
-------
A document key containing a closing bracket, or a view emitting an id
or a value string with a closing bracket won't throw off the parsing any
longer.

Change-Id: I187c551ddfe1a58c96e987b23224c7201489d677
Reviewed-on: http://review.couchbase.org/56744
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
(cherry picked from commit 54d7c5a)
Reviewed-on: http://review.couchbase.org/56747