JVMCBC-266: Reject Null, Empty or too long Key quickly.
Motivation
----------
With the current implementation if a null, empty or too long key
is passed into the core, it will be sent to the server who correctly
rejects it with INVALID_ARGUMENTS. We can do better than that.

Modifications
-------------
The locator is the first place where the key is used. The code now
checks for the correct boundaries and quickly rejects the observable
if it does not validate properly.

Tests added to verify this behavior.

Result
------
Invalid keys are quickly and semantically correctly rejected. This
also makes it easer to debug user level errors.

Change-Id: Ie11af0bb42f253e24f6e503f5e5bfd3cf9ee9821
Reviewed-on: http://review.couchbase.org/56980
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>