JVMCBC-267: Set snapshot start seqno properly on rollback
Motivation
----------
When server asks consumer to rollback, the library does not update start
snapshot sequence number in the state object, and in case latter was
less than former, the server would raise ERANGE error.

Modification
------------
Rollback sequence number have to be the same as start sequence number.

Result
------
It does not generate ERAGE errors on rollback.

Change-Id: I29e46e1e55da9b0f30102c6c98de68f2ef61800e
Reviewed-on: http://review.couchbase.org/57221
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>