JVMCBC-269: Implement Get Failover Logs (0x54) command
Motivation
----------
To initialize properly the client need to know not only sequence
numbers, but also vBucket UUIDs. One of the ways to do it is to call
0x54 command before requesting the partition stream.

Modification
------------
The change implements the aforementioned command, but also does a
refactoring required to support this command properly. The biggest part
of it, is that it is not possible to close DCP connection clearly, and
once the client opened temporary DCP connection just to fetch failover
logs, it have to close the socket, otherwise the server to do it on its
own side. The change abstracts this thing and allows to reuse DCP
connection more transparently.

Result
------
It is possible to get current state (with vBucket UUIDs).

Also fixes JVMCBC-267 and JVMCBC-268

Change-Id: I202c992a6342fd9a1e18affb300d294d49850d25
Reviewed-on: http://review.couchbase.org/57341
Well-Formed: buildbot <build@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Tested-by: Sergey Avseyev <sergey.avseyev@gmail.com>