JVMCBC-276: Add DELETE to the possible multi-mutations
Motivation:
-----------
DELETE is a valid mutation to apply in the context of multiple subdoc
mutations, yet it is missing from the enum.

Modifications:
--------------
Added DELETE to the Mutation enum. MutationCommand can now be
constructed without a fragment (Unpooled.EMPTY_BUFFER is then used),
allowing the DELETE MutationCommand to be built without providing the
unnecessary fragment part.

Added DELETE in the SubDocumentMessageTests.

Result:
-------
Fragments can be deleted via multi-mutations.

Change-Id: If4b59200fe61af55c4a79fce60373bf33257e5cc
Reviewed-on: http://review.couchbase.org/58571
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>