JVMCBC-277: Fix source relocation for netty memcache codec.
Motivation
----------
When moving from gradle to maven some part of the source shadow
relocation stopped to work subtly.

Modifications
-------------
The netty codec we ship is now properly excluded from the renaming
of the package locations. Note that this is done in a hacky way
because explicit exclusions did not do the trick (they were
ignored for some reason).

Result
------
Source jars now properly show the sources where the memcache
codec is used.

Change-Id: Ida75addcd79e4581bb641c59492e4122885f8c80
Reviewed-on: http://review.couchbase.org/58587
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>