JVMCBC-279: Add bucket name to MutationToken
Motivation
----------
The bucket name in scope of the token is needed for two reasons:

 1. tokens are scoped to a bucket and need to be identifyable for this
 2. it helps detecting problems where token for bucket A is passed
    into a query against bucket B
 3. in cross-bucket queries the code needs to be able to automatically
    scope it according to the bucket

Modifications
-------------
This changeset add the bucket string to the mutation token, makes sure it
is passed in in the handler at creation time and modifies the test suite
to also assert that they are set properly.

Result
------
the bucket name is now immutably bound to a token and can be used as a
reference scope up the stack.

Change-Id: Iccee39c55fcae8ea8064f0de302457fcea6fbe8e
Reviewed-on: http://review.couchbase.org/58841
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>