JVMCBC-282: Inform when a node is being reconnected.
Motivation
----------
We log and send events over the bus during the initial connect
as well as the terminal disconnect phase of a node. This works
as expected, but we also need to do the same during reconnect
events (like when a node is restarted).

Modifications
-------------
This changeset explicitly checks on the CONNECTING state in
the node, in addition to when the "connected" value is already
set. Those two combinations indicate a full reconnect cycled
after a previous successful connect attempt. As a result, we
inform that the node is right now disconnected and also set
the connected flag to false.

Once all the connections are established again the node goes
into CONNECTED state and another event is issued that the
node is now connected again.

Result
------
It is now visible, both from logs and over the event bus, when
a node is being reconnected (that is, disconnected and eventually
being connected again).

Change-Id: I4855844e286c80561702786fe9507d86c9a01d78
Reviewed-on: http://review.couchbase.org/59182
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>