Disable Subdoc multi-mutation support
Motivation:
-----------
The protocol for multiple mutations is likely to change before a DP
version of the server is released, so offering support for it is too
early.

Modifications:
--------------
Made classes and code related to multi mutation requests/responses not
public and/or commented.

Result:
-------
Single subdoc operations and multi lookups are still available as
experimental, but multi-mutation is not.

Change-Id: I373df3c827f48b4be14f2823316cdf4082906697
Reviewed-on: http://review.couchbase.org/59252
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>