JVMCBC-284: Don't allocate while extracting flags from get response.
Motivation
----------
There is no reason to allocate a new buffer just because we need to
extract the flags on a get response.

Modifications
-------------
The whole method call and allocation can be removed just by reading
the extra flags right away. Also, in addition the ctx doesn't need
to be passed into the method since its not needed anymore.

Result
------
Less allocations on get requests/responses.

Change-Id: I6ca30afc9e355b8ab0135a3ff8f98120262bd169
Reviewed-on: http://review.couchbase.org/59264
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>