JVMCBC-285: Cache enabled services on a per node level.
Motivation
----------
During testing it was found that the round robin node locators might
pick up the wrong node in a MDS scenario. In order for them to know
where to send the queries in the first place, it makes sense
(as an optimization) that the node itself knows which service is
currently enabled.

Modifications
-------------
Each node now remembers what services are enabled (based on an
efficient bitmap) and lookups are cheap (one volatile load) and
some simple bitshift arithmetic.

Result
------
The exposed method can be used by the locators to pick the right
nodes in the first place.

Change-Id: I97cc91a5d7580ae7fd387cb7544f5b911aedfa3c
Reviewed-on: http://review.couchbase.org/59334
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>