JVMCBC-287: Allow index based access on nodes in RequestHandler.
Motivation
----------
As a foundation for a later changeset, it must be possible to
access the nodes through an index. Currently a Set is used,
which does not provide this capability.

Modifications
-------------
The CopyOnWriteArraySet uses its list variant under the cover
anyways, so we can use it directly. Signatures and tests
have been adapted to now work with a list rather than a set,
but no other changes have been made.

Result
------
The nodes can now be accessed through an index.

Change-Id: I0db7f30784b3839318c61fefca2f5fddd6ea1305
Reviewed-on: http://review.couchbase.org/59340
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>