JVMCBC-286: Enforce usage of Java 6 APIs only
Motivation:
-----------
When developing against a recent JDK, it's not sufficient to just set
the targetVersion to 1.6, as the compiler will happily produce Java 6
bytecode that uses APIs introduced in a more recent version (eg Java 8).
This results in code that can't run on a Java 6 JRE.

The Animal-Sniffer tool has a Maven integration that allows to check for
such bad usage and prevent the build process to finish.

Modifications:
--------------
Used Animal-Sniffer Maven plugin to enforce correct usage of JDK 6 APIs
only. Exception is added to Diagnostics class by way of an annotation
since this class uses APIs only available in the Sun/Oracle implems but
in a safe guarded way.

Result:
-------
Using a too new API when developing against a recent JDK will be
detected when using Maven to build, and refused.

Change-Id: I74ae3ae22c5cf98f0016fd84c1e545435bdbcc87
Reviewed-on: http://review.couchbase.org/59339
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>