JVMCBC-289: One result / cmd on subdoc multimutate
Motivation
----------
Change to the subdocument protocol during development introduced the
notion of individual results in a multi-mutation. For now only commands
that return a value include a result (so the binary message could have
between 0 and N individual results).

Modifications
-------------
Renamed LookupResult to MultiResult and made it accept either a Lookup
or a Mutation as the operation.
This is now used by both MultiLookupResult and MultiMutationResult.

When decoding the multi mutation response, also recreate a MultiResult
for individual commands that didn't get an explicit result in the msg
so that list of commands and list of results are the same size.

Result
------
MultiMutationRequest can be made and including a COUNTER command will
have the new counter value in the response.

Change-Id: I9cda651ab64b9f86bf24c6b9fd45a53c225def74
Reviewed-on: http://review.couchbase.org/59506
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Michael Nitschinger <michael.nitschinger@couchbase.com>