JVMCBC-291: Improve ExponentialDelay documentation
Motivation:
-----------
The usage and behavior of ExponentialDelay is a bit too unclear. Eg to
achieve a delay pattern of 500ms, 1s, 2s, 4s, 4s, it is not obvious that
the growBy factor of 500 must be used rather than a lower bound of 500.

Modifications:
--------------
Added more extensive javadoc on the ExponentialDelay and Delay factory
methods.
Tests also test the attempt == 0 case.

Result:
-------
Clearer pattern of usage and examples when discovering the api of the
ExponentialDelay through javadocs.

Change-Id: I064483a5c94e1fb986571d643e81f414f3b4d846
Reviewed-on: http://review.couchbase.org/59951
Tested-by: Simon Baslé <simon@couchbase.com>
Reviewed-by: Sergey Avseyev <sergey.avseyev@gmail.com>