Disable SCRAM-SHA* by default.
Motivation
----------
Because the watson dp release has some issues regarding SCRAM,
the client code which uses it should be disabled by default.

Modifications
-------------
An undocumented property com.couchbase.scramEnabled is introduced
which is set to false and can be set to true to force using it,
mostly for internal testing going forward.

Once the server works as expected, this property will be removed
and the code is enabled all the time.

Result
------
DP testing is not blocked, but SCRAM can be enabled on demand
for further testing.

Change-Id: Ia99775c970a44fd4674b6879e43bd4b8921a6173
Reviewed-on: http://review.couchbase.org/60044
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>