JVMCBC-295: Use Yield instead of BlockingWaitStrategy for lower latency.
Motivation
----------
When investigating potential improvements in throughput and latency, it
was discovered that th default BlockingWaitStrategy on the RingBuffer is
probably not the best choice.

Modifications
-------------
The code now uses the YieldWaitStrategy which instead of using a lock
now uses Thread.yield. This also reduces GC slightly since those locks
don't need to be collected all the time.

Result
------
Slighly better performance in general.

Change-Id: Ic455723cb24a24d5428515ca7e27a47dd8258266
Reviewed-on: http://review.couchbase.org/60120
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>
Reviewed-by: Simon Baslé <simon@couchbase.com>